add workspace agent apps route
